+++ front-end/src/Components/Posts/Postapi.js
@@ -8,37 +8,18 @@ export const getAllPosts = (props) => {
   return fetch("http://172.17.28.19:5000/posts")
 };
 
-//PostAPI
+//DeleteAPI
 
-// export const postAPI = (props) => {
-//   return fetch("http://172.17.28.19:5000/posts", {
-//     method: "post",
+// export const deleteAPI = (props) => {
+//   fetch(`http://172.17.28.19:5000/posts/${props.id}`, {
+//     method: "delete",
 //     headers: {
 //       "Content-Type": "application/json",
-//       Accept: "application/json",
 //     },
-//     body: JSON.stringify({
-//       author: props.author,
-//       title: props.title,
-//       content: props.content,
-//     }),
 //   })
 //     .then((data) => data.json())
 //     .then((posts) => console.log(posts));
 // };
-
-//DeleteAPI
-
-export const deleteAPI = (props) => {
-  fetch(`http://172.17.28.19:5000/posts/${props.id}`, {
-    method: "delete",
-    headers: {
-      "Content-Type": "application/json",
-    },
-  })
-    .then((data) => data.json())
-    .then((posts) => console.log(posts));
-};
 
 //UpdateAPI
